Refactor deterministic helpers into src modules
