+++ index.html
@@ -333,6 +333,6 @@
     </div>
 
     <script src="https://cdn.jsdelivr.net/npm/quill@1.3.7/dist/quill.min.js"></script>
-    <script src="script.js"></script>
+    <script type="module" src="src/app.js"></script>
   </body>
 </html>
